pyproject.toml
[tool.poetry]
name = "codename-240"
version = "0.1.0"
description = "Simple image uploader for personal purpose"
authors = ["Jaeyeol Lee <rijgndqw012@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.6"
django-stubs = {extras = ["compatible-mypy"], version = "^4.2.6"}
requests = "^2.30.0"
django-tailwind = "^3.6.0"
django-browser-reload = "^1.12.0"
django-q2 = "^1.6.1"
blessed = "^1.20.0"
psycopg2 = "^2.9.9"
dj-database-url = "^2.1.0"
django-jinja = "^2.11.0"
jinja2 = "^3.1.3"

[tool.poetry.group.dev.dependencies]
ruff-lsp = "^0.0.53"
pyright = "^1.1.354"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.mypy]
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "core.settings"
